projects/perception/object_detection_3d/demos/voxel_object_detection_3d/requirements.txt
flask
# black
# isort>=5.7
# flake8-black
numpy
opencv-contrib-python
imutils
pandas
rplidar-roboticia